refactor svg logic out of card
